+++ qgreenland/tasks/layers/basal_thermal_state.py
@@ -1,5 +1,10 @@
 from qgreenland.tasks.common.fetch import FetchCmrGranule
-from qgreenland.tasks.common.raster import GdalEdit, GdalMDimTranslate
+from qgreenland.tasks.common.raster import (
+    BuildRasterOverviews,
+    GdalEdit,
+    GdalMDimTranslate,
+    ReformatRaster,
+)
 from qgreenland.util.luigi import LayerPipeline
 
 
@@ -21,7 +26,15 @@ class BasalThermalState(LayerPipeline):
             input_ext_override='nc',
             layer_id=self.layer_id,
         )  # ->
-        return GdalEdit(
+        gdal_edit = GdalEdit(
             requires_task=mdim_translate,
             layer_id=self.layer_id,
+        )  # ->
+        reformat = ReformatRaster(
+            requires_task=gdal_edit,
+            layer_id=self.layer_id,
+        )  # ->
+        return BuildRasterOverviews(
+            requires_task=reformat,
+            layer_id=self.layer_id
         )