+++ qgreenland/tasks/layers/netcdf_raster.py
@@ -1,6 +1,6 @@
 from qgreenland.tasks.common.fetch import FetchDataFiles
 from qgreenland.tasks.common.misc import ExtractNcDataset
-from qgreenland.tasks.common.raster import WarpRaster
+from qgreenland.tasks.common.raster import BuildRasterOverviews, WarpRaster
 from qgreenland.util.luigi import LayerPipeline
 
 
@@ -22,7 +22,11 @@ class NetCdfRaster(LayerPipeline):
             requires_task=fetch_data,
             layer_id=self.layer_id,
         )  # ->
-        return WarpRaster(
+        warp_raster = WarpRaster(
             requires_task=extract_nc_dataset,
+            layer_id=self.layer_id
+        )  # ->
+        return BuildRasterOverviews(
+            requires_task=warp_raster,
             layer_id=self.layer_id
         )