+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - New "Reference" layers:
   - QGreenland background boundary
   - QGreenland data boundary
+- Added compression to all raster layers.
+- Only use overviews on largest raster layers.
 
 # v0.53.0 (2021-02-04)